Use Frecursive_edit rather than g_run_main_loop for GTK dialogs.
The idea is that will prevent endless loops under EXWM, but it's also,
independently, a good idea (the current implementation is just a hack
that cannot go in): instead of having modal GTK dialogs, treat GTK
dialogs as input devices like the keyboard or gconf. It would be
consistent not to close down the font dialog at all, allowing the user
to try several fonts without restarting the dialog.